requirements.txt
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
pyarrow>=14.0.0
rich>=13.0.0
streamlit>=1.30.0
ccxt>=4.0.0
